fix spinner in add edit project modal
