+++ packages/client-core/src/admin/components/project/AddEditProjectModal.tsx
@@ -29,7 +29,7 @@ import { AuthState } from '@etherealengine/client-core/src/user/services/AuthSer
 import { DefaultUpdateSchedule } from '@etherealengine/common/src/interfaces/ProjectPackageJsonType'
 import { ProjectBranchType, ProjectCommitType, ProjectType } from '@etherealengine/common/src/schema.type.module'
 import { toDateTimeSql } from '@etherealengine/common/src/utils/datetime-sql'
-import { getMutableState } from '@etherealengine/hyperflux'
+import { NO_PROXY, getMutableState } from '@etherealengine/hyperflux'
 import Button from '@etherealengine/ui/src/primitives/tailwind/Button'
 import Input from '@etherealengine/ui/src/primitives/tailwind/Input'
 import Label from '@etherealengine/ui/src/primitives/tailwind/Label'
@@ -45,6 +45,7 @@ import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { CiCircleCheck, CiCircleRemove, CiWarning } from 'react-icons/ci'
 import { HiMiniClipboardDocumentList } from 'react-icons/hi2'
+import { NotificationService } from '../../../common/services/NotificationService'
 import { ProjectUpdateService, ProjectUpdateState } from '../../services/ProjectUpdateService'
 
 const autoUpdateIntervalOptions = [
@@ -91,19 +92,10 @@ const getTempProject = () => ({
   commitDate: toDateTimeSql(new Date())
 })
 
-export default function AddEditProjectModal({
-  update,
-  inputProject,
-  onSubmit,
-  processing
-}: {
-  update: boolean
-  inputProject?: ProjectType
-  onSubmit: () => void
-  processing: boolean
-}) {
+export default function AddEditProjectModal({ update, inputProject }: { update: boolean; inputProject?: ProjectType }) {
   const { t } = useTranslation()
   const showAutoUpdateOptions = useHookstate(false)
+  const modalProcessing = useHookstate(false)
 
   const project = update && inputProject ? inputProject : getTempProject()
 
@@ -372,6 +364,28 @@ export default function AddEditProjectModal({
     }
   }, [projectUpdateStatus?.value?.triggerSetDestination])
 
+  const handleSubmit = async () => {
+    modalProcessing.set(true)
+    const projectUpdateStatus = getMutableState(ProjectUpdateState)['tempProject'].get(NO_PROXY)
+
+    try {
+      await ProjectService.uploadProject({
+        sourceURL: projectUpdateStatus.sourceURL,
+        destinationURL: projectUpdateStatus.destinationURL,
+        name: projectUpdateStatus.projectName,
+        reset: true,
+        commitSHA: projectUpdateStatus.selectedSHA,
+        sourceBranch: projectUpdateStatus.selectedBranch,
+        updateType: projectUpdateStatus.updateType,
+        updateSchedule: projectUpdateStatus.updateSchedule
+      })
+      PopoverState.hidePopupover()
+    } catch (err) {
+      NotificationService.dispatchNotify(err.message, { variant: 'error' })
+    }
+    modalProcessing.set(false)
+  }
+
   return (
     <Modal
       className="relative max-h-full w-[50vw] max-w-2xl p-4"
@@ -380,8 +394,8 @@ export default function AddEditProjectModal({
         ProjectUpdateService.clearProjectUpdate(project.name)
         PopoverState.hidePopupover()
       }}
-      onSubmit={onSubmit}
-      submitLoading={processing}
+      onSubmit={handleSubmit}
+      submitLoading={modalProcessing.value}
     >
       <div className="grid gap-6">
         <div className="grid gap-2">
