+++ packages/client-core/src/admin/components/project/ProjectTopMenu.tsx
@@ -23,25 +23,22 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { NotificationService } from '@etherealengine/client-core/src/common/services/NotificationService'
 import { PopoverState } from '@etherealengine/client-core/src/common/services/PopoverState'
 import { ProjectService, ProjectState } from '@etherealengine/client-core/src/common/services/ProjectService'
 import config from '@etherealengine/common/src/config'
-import { NO_PROXY, getMutableState, useHookstate } from '@etherealengine/hyperflux'
+import { getMutableState, useHookstate } from '@etherealengine/hyperflux'
 import Button from '@etherealengine/ui/src/primitives/tailwind/Button'
 import LoadingView from '@etherealengine/ui/src/primitives/tailwind/LoadingView'
 import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { HiArrowPath, HiPlus } from 'react-icons/hi2'
 import { AuthState } from '../../../user/services/AuthService'
-import { ProjectUpdateState } from '../../services/ProjectUpdateService'
 import AddEditProjectModal from './AddEditProjectModal'
 import UpdateEngineModal from './UpdateEngineModal'
 
 export default function ProjectTopMenu() {
   const { t } = useTranslation()
   const projectState = useHookstate(getMutableState(ProjectState))
-  const modalProcessing = useHookstate(false)
 
   ProjectService.useAPIListeners()
 
@@ -59,28 +56,6 @@ export default function ProjectTopMenu() {
       if (interval) clearInterval(interval)
     }
   }, [projectState.rebuilding.value])
-
-  const handleSubmit = async () => {
-    modalProcessing.set(true)
-    const projectUpdateStatus = getMutableState(ProjectUpdateState)['tempProject'].get(NO_PROXY)
-
-    try {
-      await ProjectService.uploadProject({
-        sourceURL: projectUpdateStatus.sourceURL,
-        destinationURL: projectUpdateStatus.destinationURL,
-        name: projectUpdateStatus.projectName,
-        reset: true,
-        commitSHA: projectUpdateStatus.selectedSHA,
-        sourceBranch: projectUpdateStatus.selectedBranch,
-        updateType: projectUpdateStatus.updateType,
-        updateSchedule: projectUpdateStatus.updateSchedule
-      })
-      PopoverState.hidePopupover()
-    } catch (err) {
-      NotificationService.dispatchNotify(err.message, { variant: 'error' })
-    }
-    modalProcessing.set(false)
-  }
 
   const authState = useHookstate(getMutableState(AuthState))
   const user = authState.user
@@ -132,9 +107,7 @@ export default function ProjectTopMenu() {
           startIcon={<HiPlus />}
           size="small"
           onClick={() => {
-            PopoverState.showPopupover(
-              <AddEditProjectModal processing={modalProcessing.value} onSubmit={handleSubmit} update={false} />
-            )
+            PopoverState.showPopupover(<AddEditProjectModal update={false} />)
           }}
         >
           {t('admin:components.project.addProject')}
